requirements.txt
Flask
fpdf
pdfplumber
python-docx
euriai
werkzeug
python-dotenv